pyproject.toml
[tool.poetry]
name = "ot-projekti"
version = "0.1.0"
description = ""
authors = ["meri saynatkari <meri.saynatkari@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pygame = "^2.0.1"
invoke = "^1.5.0"
python-dotenv = "^0.17.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"
coverage = "^5.5"
pylint = "^2.7.4"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
